+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Master
+* add brain structures
+* add liver vessels
 
 
 ## Release 2.3.0